Implement Periodic Tick Predicates
Replaced some calls to the various periodic scoreboard objectives with the new periodic_tick predicate. Objective bp.tick was completely replaced (except for when it was being used for something else), while the others were replaced only when calls looked for an objective value of 1 (leaving the staggered calls was a good idea).