Fix pax objective conflict
Previous "bp.pax" objective was preventing the "bp.pax" trigger from being instantiated correctly, and since it was constantly being set up to 10 when the player was in pax, the game was misinterpreting that as the player triggering the option to set pax as their home dimension.

I've removed the original "bp.pax" objective entirely, as it wasn't actually used for anything!